requirements.txt
agno==1.6.2
duckduckgo_search==8.0.4
easyocr==1.7.2
fpdf==1.7.2
groq==0.28.0
huggingface-hub==0.27.0
humanize==4.12.0
nltk==3.9.1
nncf==2.17.0
openvino==2025.1.0
openvino-genai==2025.1.0.0
openvino-telemetry==2025.1.0
openvino-tokenizers==2025.1.0.0
optimum==1.26.1
optimum-intel==1.24.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
pipx==1.7.1
PyAudio==0.2.14
PyMuPDF==1.26.0
pymupdf4llm==0.0.24
pypdf==5.1.0
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.0.1
pyttsx3==2.98
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
selenium==4.28.1
setuptools==74.1.3
SpeechRecognition==3.14.3
streamlit==1.46.1
tokenizers==0.21.0